+++ src/windows/mod.rs
@@ -3,6 +3,7 @@ use winit::event_loop::EventLoop;
 use crate::windows::app::{App, AppMessage};
 
 mod app;
+mod context_menu;
 mod egui_glue;
 mod message_window;
 mod taskbar;