+++ src/windows/windows/mod.rs
@@ -1,2 +1,2 @@
-mod resize;
+mod config_window;
 mod switcher;